can add agents to specific locations and map_layout is no its own object
